+++ R/BARIS_explain.R
@@ -21,10 +21,15 @@ BARIS_explain <- function(datasetId) {
 
   checkmate::assert_character(datasetId)
 
-
   basic_url <- "https://www.data.gouv.fr/api/1/datasets/"
 
   final_url <- paste("https://www.data.gouv.fr/api/1/datasets/", datasetId, "/", sep = "")
+
+  response_status <- is_response_successfull(final_url)
+
+  if (!response_status) {
+    return(NULL)
+  }
 
   # set memoise cache fct
   mem_fromJSON <- BARIS_set_fromJSON()