+++ NEWS.md
@@ -1,11 +1,17 @@
+# BARIS 1.1.2
+* The functions that query data now check for the API response, and if not successful, they return `NULL`.
+
+* In `BARIS_search`, removing the `n_pages` parameter and adding two new parameters to query the results, namely: `page_number` and `page_size`
+
+
 # BARIS 1.1.1
 
 * minor bugs fixing 
-* added assertion functionalities
+* added assertion features
 
 # BARIS 1.1.0
 
-* I've and add-in that allows the user to use the package in an interactive manner. 
+* I've added an add-in that allows the user to use the package in an interactive manner. 
 
 # BARIS 1.0.0